# packages/kanban-viewer/openapi.yaml
openapi: 3.1.0
info:
  title: A(i)-Team Kanban Viewer API
  version: 1.0.0
  description: |
    REST API for the A(i)-Team Kanban Viewer. Manages work items, missions,
    agent lifecycle, and real-time board events for the TDD pipeline.

    ## Project Isolation
    All requests (except `/api/projects`) require the `X-Project-ID` header.
    Projects are created automatically on first use via `POST /api/projects`.

    ## Error Format
    All errors return a consistent envelope:
    ```json
    { "success": false, "error": { "code": "ERROR_CODE", "message": "...", "details": {} } }
    ```
servers:
  - url: http://localhost:3000
    description: Local development

# ─── Reusable components ─────────────────────────────────────────────────────
components:
  parameters:
    ProjectId:
      name: X-Project-ID
      in: header
      required: true
      description: Project identifier for multi-tenant isolation
      schema:
        type: string
        example: my-project
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - INVALID_TRANSITION
                - WIP_LIMIT_EXCEEDED
                - INVALID_STAGE
                - DEPENDENCY_CYCLE
                - DEPENDENCIES_NOT_MET
                - OUTPUT_COLLISION
                - CLAIM_CONFLICT
                - ITEM_CLAIMED
                - NOT_CLAIMED
                - CLAIM_MISMATCH
                - ITEM_NOT_FOUND
                - MISSION_NOT_FOUND
                - NO_ACTIVE_MISSION
                - INVALID_MISSION_STATE
                - PROJECT_NOT_FOUND
                - STAGE_NOT_FOUND
                - CONFLICT
                - UNAUTHORIZED
                - SERVER_ERROR
                - DATABASE_ERROR
            message:
              type: string
            details:
              description: Optional structured context for the error
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /api/tuning/proposals:
    parameters:
      - $ref: '#/components/parameters/ProjectId'

    post:
      operationId: createTuningProposal
      summary: Create a TuningProposal for corroborated fingerprint(s)
      description: |
        Creates the proposal directly with status='accepted' — the POST
        itself is the agreement (Phase A: no eager browsing draft). Rejected
        with 422 NOT_CORROBORATED unless every linked fingerprint has
        `distinctMissions >= 3` globally.
      tags: [tuning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fingerprints, targetSurface, altitude, proposalText]
              properties:
                fingerprints:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  description: Fingerprint slugs this proposal agrees to act on
                targetSurface:
                  type: string
                  example: 'agents/ba.md'
                altitude:
                  type: string
                  enum: [skill-text, agent-prompt, hook]
                proposalText:
                  type: string
      responses:
        '201':
          description: Proposal created (status='accepted')
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    required: [id, status, fingerprints]
                    properties:
                      id:
                        type: integer
                      status:
                        type: string
                        enum: [accepted]
                      fingerprints:
                        type: array
                        items:
                          type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: NOT_CORROBORATED — not every linked fingerprint has distinctMissions >= 3
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
...